test: cli - extract core logic
As part of the test framework cleanup, we want a better way to isolate the bugs by having a nice UX towards mimicking user (or llm) actions.

Right now, the cli commands were copy-pasted, which adds some maintenance burden -- which we want to avoid. Had to do it for some sub-functions running git directly (which we don't want to be interactive)

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>